backend/requirements.txt
clerk-backend-api>=3.3.1
fastapi>=0.117.1
google-generativeai>=0.8.5
openai>=1.109.1
python-dotenv>=1.1.1
sqlalchemy>=2.0.43
svix>=1.76.1
uvicorn>=0.37.0